Extract SQL graph contract helpers partial
